update tests for reducers
